Improve unit tests and tests structure
Made an improvement to the test for the raw nmap command, by executing
'nmap --version' instead of just 'nmap'. This results in an exit code of
0, instead of an error exit code.
Moved the unit tests to different files to represent the command line
structure, this will scale better when more tests and commands get
added. Renamed of the the unit tests because of this.