Add '--subnet' option to 'raw nmap' command
This options appends the IPv4 subnet associated with the interface as
target. Suppose the interface 'eth0' has IP address '192.168.1.60' and
netmask '255.255.255.0'. The option '--iface-subnet eth0' will append
the nmap command with the target '192.168.1.0/24'.

This initial implementation is functional, but has a lot of room for
improvement. Various TODOs are added and edge cases should be handled.

This functionality uses the 'netifaces' package, which is added to the
requirements in setup.py.