Fix docstyle errors in entire project
Multiple minor fixes in order to conform to docstyle. One of these,
altough not reported by pydocstyle, is the newline at the start of a
docstring. Doing the following

	"""
	Start of docstring

will actually instert a newline at the start of the docstring, resulting
in "\nStart of docstring". Therefore, also with multiline docstrings,
start the first line directly after the quotes.