Rewrite turret.core.util.interface_subnets()
This new implementation returns a set of all the IPv4 subnets associated
with an interface. It does not implement IPv6 yet, but a TODO for this
is included. The function is now also annotated with types.

Instead of netifaces, psutil is now used. This change is also reflected
in the setup.py, where netifaces is replaced with psutil.