Add 'capture_output' option to extenders.Program
The 'capture_output' option specifies if the stdout and stderr of the
Program should be captured and stored in the 'output' attribute of
CompletedProgram. If the output is captured, it will no longer be
displayed by default, so it is set to False in raw Nmap. The default
value is True.